[function] 기본적인 repost 운영자 화면 구현

+++ src/main/java/store/babel/babel/domain/report/dto/ReportSearchRequest.java
@@ -22,7 +22,7 @@ public class ReportSearchRequest
 
     @Min(0)
     @Max(100)
-    private int size;
+    private Long size;
 
     private ReportSortType sortType;
